# Sql summary optimisation phase 2