--- a/test/OpenTelemetry.Contrib.Shared.Tests/OpenTelemetry.Contrib.Shared.Tests.csproj
+++ b/test/OpenTelemetry.Contrib.Shared.Tests/OpenTelemetry.Contrib.Shared.Tests.csproj
@@ -32,6 +32,7 @@
 
   <ItemGroup>
     <EmbeddedResource Include="SqlProcessorTestCases.json" LogicalName="SqlProcessorTestCases.json" />
+    <EmbeddedResource Include="SqlProcessorAdditionalTestCases.json" LogicalName="SqlProcessorAdditionalTestCases.json" />
   </ItemGroup>
 
 </Project>