@stevejgordon 2025-09-15
A couple of the tests are failin gon net462 - I'll look into those tomorrow.