@stevejgordon 2025-09-18
@martincostello The EF Core tests pass on my machine. Are you able to rerun these please?