@stevejgordon 2025-09-30
@martincostello I have one more (hopefully) thing to tackle which is the truncation once we decide on option a) or b).